gestor_tareas/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib
jinja2
